LayoutTests/fast/canvas/webgl/index-validation-expected.txt

Tests that index validation verifies the correct number of indices

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing with valid indices
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
PASS gl.getError() is 0
PASS gl.drawElements(gl.TRIANGLES, 3, gl.UNSIGNED_SHORT, 0) is undefined.
PASS gl.getError() is 0
Testing with out-of-range indices
Enable vertices, valid
PASS gl.getError() is 0
PASS gl.drawElements(gl.TRIANGLES, 3, gl.UNSIGNED_SHORT, 0) is undefined.
PASS gl.getError() is 0
Enable normals, out-of-range
PASS gl.getError() is 0
PASS gl.drawElements(gl.TRIANGLES, 3, gl.UNSIGNED_SHORT, 0) is undefined.
PASS gl.getError() is gl.INVALID_OPERATION
Test with enabled attribute that does not belong to current program
Enable an extra attribute with null
PASS gl.getError() is 0
PASS gl.drawElements(gl.TRIANGLES, 3, gl.UNSIGNED_SHORT, 0) is undefined.
PASS gl.getError() is gl.INVALID_OPERATION
Enable an extra attribute with insufficient data buffer
PASS gl.getError() is 0
PASS gl.drawElements(gl.TRIANGLES, 3, gl.UNSIGNED_SHORT, 0) is undefined.
PASS gl.getError() is 0
PASS successfullyParsed is true

TEST COMPLETE
